indent

+++ City-Zen/src/engine/managers/SatisfactionManager.java
@@ -11,19 +11,15 @@ import gui_data.GuiConstants;
 
 public final class SatisfactionManager 
 {
-	public int mathSatisfaction() 
-	{
-	District[][] map = Game.getINSTANCE().getDistrictMap();
-	int satisfaction;
-	satisfaction = 0;
-	RailSquare railSquare[][] = new RailSquare[GuiConstants.SQUARE_PER_COLUMN][GuiConstants.SQUARE_PER_ROW];
-	int column = 0,row = 0;
-		while( column<GuiConstants.SQUARE_PER_ROW) 
-		{
-			while(row<GuiConstants.SQUARE_PER_COLUMN) 
-			{
-				if (map[column][row].getType() == 2)
-				{
+	public double mathSatisfaction(){
+		District[][] map = Game.getINSTANCE().getDistrictMap();
+		double satisfaction;
+		satisfaction = 0;
+		RailSquare railSquare[][] = new RailSquare[GuiConstants.SQUARE_PER_COLUMN][GuiConstants.SQUARE_PER_ROW];
+		int column = 0,row = 0;
+		while( column<GuiConstants.SQUARE_PER_ROW){
+			while(row<GuiConstants.SQUARE_PER_COLUMN){
+				if (map[column][row].getType() == 2){
 					satisfaction = satisfaction + map[column][row].getSatisfaction()*((Residencial)map[column][row]).getNbHab();
 				}
 				row++;
